aws sdk 관련 코드 하나로 합침, 함수 util 모듈로 뺌
