+++ src/aws_iam_sync.py
@@ -5,6 +5,7 @@ import boto3
 from fastapi import HTTPException
 from pymongo import MongoClient
 
+from . import util
 from .config import conf
 
 
@@ -21,59 +22,13 @@ class awsIamSync:
         self.awsPolicies_collection = self.db["awsPolicies"]
         self.awsUsers_collection = self.db["awsUsers"]
 
-    def remove_special_characters(self, string):
-        return re.sub(r"[^A-Za-z0-9\s]", "", string)
-
-    # 문자열의 문자 수를 반환해준다.
-    def count_string(self, input):
-        string = str(input)
-        string = string.replace(" ", "")
-        return len(string)
-ㄷ
-    # 정책을 합칠 수 있는 조합을 반환한다.
-    def calculate_set(self, string_lengths):
-        limit_string = 5000  # 최대 문자 수
-        string_lengths_copy = string_lengths.copy()  # 입력받은 문자 수 리스트를 복사한다.
-        string_lengths_copy.sort()  # 오름차 순으로 정렬
-        result = []  # 결과를 저장할 리스트
-        while len(string_lengths_copy) > 1:  # list안의 요소가 1개 이상일 경우 반복한다.
-            index_set = []  # 찾아낸 조합을 저장할 리스트
-            max_value = string_lengths_copy[-1]
-            min_value = string_lengths_copy[0]
-            max_value_index = string_lengths.index(
-                string_lengths_copy.pop(-1)
-            )  # 최댓값의 인덱스를 index_set 리스트에 추가하고 pop
-            index_set.append(max_value_index)
-            string_lengths[max_value_index] = -1  # 사용한 숫자를 재사용하지 못하게 -1로 변경
-            total_value = max_value + min_value
-            while total_value < limit_string:  # 최대 문자 수보다 작을 경우에만 반복
-                min_value_index = string_lengths.index(
-                    string_lengths_copy.pop(0)
-                )  # 최솟값의 인덱스를 index_set에 추가하고 pop
-                index_set.append(min_value_index)
-                string_lengths[min_value_index] = -1
-                if string_lengths_copy:  # 남은 값이 있을 경우 리스트의 최솟값을 다시 할당
-                    min_value = string_lengths_copy[0]
-                    total_value += min_value
-                else:
-                    break
-            if len(index_set) > 1:
-                result.append(index_set)
-            else:
-                result.extend(index_set)
-        else:
-            if string_lengths_copy:  # 반복문이 끝나고 값이 하나 있을 경우에 마지막 값의 index를 result에 추가
-                result.append(string_lengths.index(string_lengths_copy[0]))
-
-        return result
-
     # 정책을 합쳐준다.
     def policy_compress(self, policies):
         result_docs_list = []
         string_length = []
         for policy in policies:
-            string_length.append(self.count_string(policy["Statement"]))
-        index_set = self.calculate_set(string_length)
+            string_length.append(util.count_string(policy["Statement"]))
+        index_set = util.calculate_set(string_length)
         for set in index_set:
             if isinstance(set, int):
                 result_docs_list.append(set)
@@ -89,14 +44,14 @@ class awsIamSync:
         serach_string = "arn:aws:iam::aws:policy/"  # aws 관리형 signature
         docs_list = []
         for policy_arn in policy_arn_list:
-            arn, policy_name = map(str, policy_arn.split("/"))
+            _, policy_name = map(str, policy_arn.split("/"))
             if serach_string in policy_arn:
                 collection = self.db["awsPolicies"]
                 query_result = collection.find_one({"Arn": policy_arn})
                 for index, statement in enumerate(
                     query_result["Document"]["Statement"]
-                ):ㅇ
-                    statement["Sid"] = self.remove_special_characters(
+                ):
+                    statement["Sid"] = util.remove_special_characters(
                         f"{policy_name}{index+1}"
                     )
                 docs_list.append(query_result["Document"])
@@ -106,7 +61,7 @@ class awsIamSync:
                 for index, statement in enumerate(
                     query_result["Document"]["Statement"]
                 ):
-                    statement["Sid"] = self.remove_special_characters(
+                    statement["Sid"] = util.remove_special_characters(
                         f"{policy_name}{index+1}"
                     )
                 docs_list.append(query_result["Document"])
@@ -185,3 +140,23 @@ class awsIamSync:
             self.delete_old_position(aws_user_name, postion_name)
         for postion_name in new_user_data.attachedPosition:
             self.position_sync_aws(new_user_data.awsAccount, postion_name)
+
+    def create_group(self, group_name):
+        return self.aws_sdk.create_group(GroupName=group_name)
+
+    def add_user_to_group(self, group_name, user_name):
+        return self.aws_sdk.add_user_to_group(UserName=user_name, GroupName=group_name)
+
+    def attach_group_policy(self, group_name, policy_arn):
+        return self.aws_sdk.attach_group_policy(
+            GroupName=group_name, PolicyArn=policy_arn
+        )
+
+    def attach_group_position(group_name, position_id):
+        pass
+
+    def update_awsUsers(user_name):
+        pass
+
+    def update_awsGroups(group_name):
+        pass
